% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_and_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw NOAA dataset}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{data frame of raw NOAA data}
}
\value{
Returns a dataframe of the clean NOAA data
}
\description{
Clean raw NOAA dataset
}
\examples{
# Read in raw NOAA data
          \dontrun{raw_data <- eq_read_data("./inst/extdata/earthquakes.tsv")}
          \dontrun{clean_data <- eq_clean_data(raw_data)}

}
