% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_and_clean_data.R
\name{eq_country_clean}
\alias{eq_country_clean}
\title{Populate country column}
\usage{
eq_country_clean(location)
}
\arguments{
\item{location}{The country column in the NOAA dataset}
}
\value{
string "US" if location is a American state
}
\description{
Returns "US" for any
}
\examples{
\dontrun{eq_country_clean("UTAH")}

}
