% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Make annotation text for eq_map()}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Data frame of earthquake observations to be labeled}
}
\value{
A string in HTML style to be used as annotation text
}
\description{
Constructs label column to be used as the annotation text in the leaflet map
}
\details{
Skips over labels for missing values
}
\examples{
\dontrun{eq_create_label(data)}

}
