% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_and_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean NOAA Location Name Column}
\usage{
eq_location_clean(location_name)
}
\arguments{
\item{location_name}{the location name column in the NOAA dataset}
}
\value{
the cleaned location name column for the NOAA dataset
}
\description{
Clean NOAA Location Name Column
}
\details{
Strips out the country name (and the colon) from the location name column and adds title casing
}
\examples{
\dontrun{eq_location_clean("GREECE:  MOUNT TAYGETUS")}

}
