% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquake events from NOAA dataset}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Data frame of the cleaned NOAA earthquake dataset}

\item{annot_col}{String of name of column to be used for point annotations}
}
\value{
An interactive leaflet map with points of earthquake observations
}
\description{
Plots earthquake observations to an interactive where the points are the coordinates of the epicenters of the earthquakes.
}
\examples{
# Read in raw NOAA data
          \dontrun{raw_data <- eq_read_data("./inst/extdata/earthquakes.tsv")}
          \dontrun{clean_data <- eq_clean_data(raw_data)}
          \dontrun{min_date <- ymd("2020-01-01")}
          \dontrun{max_date <- ymd("2020-01-31")}
          \dontrun{data <- filter(clean_data, date >= min_date & date <= max_date)}
          \dontrun{data \%>\%}
          \dontrun{eq_map(annot_col="date")}

}
