% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_and_clean_data.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Read in NOAA dataset}
\usage{
eq_read_data(filename)
}
\arguments{
\item{filename}{character string of path to raw NOAA data file}
}
\value{
Returns a dataframe of the raw NOAA data
}
\description{
Read in NOAA dataset
}
\details{
throws an error if file path does not exist
}
\examples{
\dontrun{raw_data <- eq_read_data("./inst/extdata/earthquakes.tsv")}

}
