% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create geom_timeline Layer}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{An aesthetic mapping}

\item{data}{A data frame of earthquake observations}

\item{stat}{A string of the stat applied to the data for plot layer}

\item{position}{A string of the position adjustment}

\item{na.rm}{A logical to remove NAs from data}

\item{show.legend}{A logical to display the legend}

\item{inherit.aes}{A logical to override the default aesthetics}

\item{...}{Other arguments passed on to methods.}
}
\value{
a ggproto layer to plot a timeline of earthquakes with their magnitude and death toll
}
\description{
Plots a timeline of earthquake events in the NOAA dataset with
             size and color of plotted points indicating information about
             each earthquake.
}
\examples{
# Read in raw NOAA data
          \dontrun{raw_data <- eq_read_data("./inst/extdata/earthquakes.tsv")}
          \dontrun{clean_data <- eq_clean_data(raw_data)}
          \dontrun{min_date <- ymd("2020-01-01")}
          \dontrun{max_date <- ymd("2020-01-31")}
          \dontrun{data <- filter(clean_data, date >= min_date & date <= max_date)}
          \dontrun{ggplot(data, aes(x=date, size=magnitude, fill=deaths, color=deaths,)) +
                   geom_timeline()}

}
\seealso{
GeomTimeLine
}
