% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Create geom_timeline_label Layer}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{An aesthetic mapping}

\item{data}{A data frame of earthquake observations}

\item{stat}{A string of the stat applied to the data for plot layer}

\item{position}{A string of the position adjustment}

\item{na.rm}{A logical to remove NAs from data}

\item{show.legend}{A logical to display the legend}

\item{inherit.aes}{A logical to override the default aesthetics}

\item{...}{Other arguments passed on to methods.}
}
\value{
a ggproto layer to plot labels of top earthquakes by magnitude
}
\description{
Create geom_timeline_label Layer
}
\examples{
# Read in raw NOAA data
         \dontrun{raw_data <- eq_read_data("./inst/extdata/earthquakes.tsv")}
         \dontrun{clean_data <- eq_clean_data(raw_data)}
         \dontrun{min_date <- ymd("2020-01-01")}
         \dontrun{max_date <- ymd("2020-01-31")}
         \dontrun{data <- filter(clean_data, date >= min_date & date <= max_date)}
         \dontrun{ggplot(data, aes(x=date, label=location_name,
                                   size=magnitude, fill=deaths,
                                   color=deaths, n_max=5)) +
                  geom_timeline_label()}

}
\seealso{
GeomTimeLine
}
