% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{make_label}
\alias{make_label}
\title{Make labels for each earthquake attribute}
\usage{
make_label(label_name, column)
}
\arguments{
\item{label_name}{String of a label name for column}

\item{column}{Data frame column from earthquake data to be labeled}
}
\value{
A string in HTML style to be used as annotation text
}
\description{
Constructs individual labels for a given earthquake attribute
}
\details{
Skips over labels for missing values
}
\examples{
\dontrun{make_label("Location", data$location_name)}

}
